+++ test/specs/record_type_test.dart
@@ -5,12 +5,10 @@
 import 'package:code_builder/code_builder.dart';
 import 'package:test/test.dart';
 
-import '../common.dart';
+import '../matchers.dart';
 
 void main() {
   late DartEmitter emitter;
-
-  useDartfmt();
 
   setUp(() => emitter = DartEmitter.scoped(useNullSafetySyntax: true));